// manual-scripts/shared/src/types.ts
export interface CollectionMetaData {
    name?: string;
    symbol?: string;
    elementCount?: number;
    updateAuthority?: string;
    creator?: string;
    description?: string;
    family?: string;
    image?: string;
}

export interface CollectionAttributeData {
    attributes: CollectionAttribute[];
}

export interface CollectionPriceData {
    volume: CollectionVolume;
    floorPrice: CollectionPrice;
    ceilingPrice: CollectionPrice;
}

export interface CollectionAttribute {
    name?: string;
    key?: string;
    values: CollectionAttributeValue[];
}
// ...
export enum EntityType {
    BidTransaction = 'bidTransaction',
    AskTransaction = 'askTransaction',
    CancelBidTransaction = 'cancelBidTransaction',
    CancelAskTransaction = 'cancelAskTransaction',
    ExecuteSaleTransaction = 'executeSaleTransaction',
    MarketplaceTransactionForNFT = 'marketplaceTransaction4NFT',
    MarketplaceTransactionForWallet = 'marketplaceTransaction4Wallet',

    BidOrderState = 'bidOrderState',
    AskOrderState = 'askOrderState',

    UserRefreshTokenEntity = 'userRefreshToken',
    NonceEntity = 'nonce',

    MetaplexOnChainMetadata = 'MetaplexOnChainMetadata',
    MintAddress = 'mintAddress',
    NFTOffchainMetadata = 'NFToffchainMetadata',
    MetaplexOffchainMetadata = 'MetaplexOffchainMetadata',
    Trigger = 'trigger',
    CensorTag = "censorTag",
    CollectionID = "collectionID",
    UserFlag = "userFlag",
    StaleFlag = "staleFlag",

    CollectionAttributeData = 'collectionAttributeData',
    CollectionMetaData = 'collectionMetaData',
    CollectionPriceData = 'collectionPriceData',

    NFT4Collection = 'NFT4Collection',
    Collection4NFT = 'collection4NFT',
    PlaydustCollection = 'playdustCollection',
    Transaction = 'transaction',

    MarketData4Collection = 'marketData4Collection',
    OffchainMetadata4Collection = 'offchainMetadata4Collection',
    NFTRarity4Collection = 'NFTRarity4Collection',

    NFTRarity = 'NFTRarity',
}
// ...
export type StringPublicKey = string;
// ...
export class MintEntityData{
    decimals: number;
    freezeAuthority: StringPublicKey;
    isInitialized: boolean;
    mintAuthority: StringPublicKey;
    supply: number;
    constructor( args: {
        decimals: number;
        freezeAuthority: StringPublicKey;
        isInitialized: boolean;
        mintAuthority: StringPublicKey;
        supply: number;
    }){
        this.decimals = args.decimals;
        this.freezeAuthority = args.freezeAuthority;
        this.isInitialized = args.isInitialized;
        this.mintAuthority = args.mintAuthority;
        this.supply = args.supply;
    }
}

export interface PlaydustCollectionData {
    type: CollectionType;
    id: string;
}

export enum CollectionType {
    Metaplex = 'Metaplex',
    Manual = 'Manual',
    MagicEden = 'MagicEden',
    Derived = 'Derived',
}
export interface Entity {
    globalId: string;
    id?: string;
    type?: EntityType;
    primaryEntity?: string;
    createdAt?: Date;
    updatedAt: Date;
    // FIXME: Use a generic type here
    data?: any;
};
// ...
export interface Transaction {
    /** The transaction message */
    message: Message;
    /** The transaction signatures */
    signatures: string[];
}

interface ByMarketplaceCollectionData<T> {
    byMarketplace?: Map<Marketplace, T>;
    global: T;
};

export interface VolumeData {
    d1: number;
    d7: number;
    d30: number;
    total: number;
};

export type CollectionPrice = ByMarketplaceCollectionData<number>;
export type CollectionVolume = ByMarketplaceCollectionData<VolumeData>;